[Refactor] Reminder VC들에 공통적으로 들어가는 요소들을 갖는 BaseReminderVC 구성
