+++ Favor/Favor/Sources/Scenes/Reminder/Reactor/EventDetailViewReactor.swift
@@ -1,0 +1,38 @@
+//
+//  EventDetailViewReactor.swift
+//  Favor
+//
+//  Created by 이창준 on 2023/04/02.
+//
+
+import ReactorKit
+
+final class EventDetailViewReactor: Reactor {
+
+  // MARK: - Properties
+
+  var initialState: State
+
+  enum Action {
+
+  }
+
+  enum Mutation {
+
+  }
+
+  struct State {
+
+  }
+
+  // MARK: - Initializer
+
+  init() {
+    self.initialState = State()
+  }
+
+
+  // MARK: - Functions
+
+
+}
