+++ Favor/Favor/Sources/Scenes/Reminder/ViewControllers/EventDetailVC.swift
@@ -1,0 +1,37 @@
+//
+//  EventDetailVC.swift
+//  Favor
+//
+//  Created by 이창준 on 2023/04/02.
+//
+
+import UIKit
+
+import FavorKit
+import ReactorKit
+import SnapKit
+
+final class EventDetailViewController: BaseViewController, View {
+
+  // MARK: - Constants
+
+  // MARK: - Properties
+
+  // MARK: - UI Components
+
+  // MARK: - Life Cycle
+
+  // MARK: - Binding
+
+  func bind(reactor: EventDetailViewReactor) {
+    // Action
+
+    // State
+
+  }
+
+  // MARK: - Functions
+
+  // MARK: - UI Setups
+
+}
