+++ Favor/FavorKit/Sources/FavorKit/Extensions/Date+.swift
@@ -16,6 +16,12 @@ extension Date {
     Int(self.toMonthString()) ?? 0
   }
 
+  public func toString() -> String {
+    let formatter = DateFormatter()
+    formatter.dateFormat = "yyyy년 M월 d일"
+    return formatter.string(from: self)
+  }
+
   public func toYearString() -> String {
     let formatter = DateFormatter()
     formatter.dateFormat = "yyyy"
